Move response parsers
