--- src/avalan/agent/orchestrator/response/parsers/__init__.py
@@ -1,7 +1,0 @@
-from typing import Any, Iterable, Protocol
-
-
-class StreamParser(Protocol):
-    async def push(self, token_str: str) -> Iterable[Any]: ...
-
-    async def flush(self) -> Iterable[Any]: ...
